# fix(2024-06-01): use registered domain for third parties to count sub-domains