--- a/sql/all-observed-domains-query.sql
+++ b/sql/all-observed-domains-query.sql
@@ -1,20 +1,9 @@
-SELECT
-  domain,
-  COUNT(0) AS totalOccurrences
+SELECT domain, COUNT(*) as totalOccurrences
 FROM (
-  SELECT
-    page,
-    NET.HOST(url) AS domain,
-    COUNT(0) AS totalOccurrences
-  FROM
-    `httparchive.requests.2022_01_01_mobile`
-  GROUP BY
-    page,
-    domain
+  SELECT page, NET.REG_DOMAIN(url) AS domain,
+  FROM `httparchive.requests.2024_06_01_mobile`
+  GROUP BY page, domain
 )
-GROUP BY
-  domain
-HAVING
-  totalOccurrences >= 50
-ORDER BY
-  totalOccurrences DESC
+GROUP BY domain
+HAVING totalOccurrences >= 50
+ORDER BY totalOccurrences DESC