--- a/sql/entity-per-page.sql
+++ b/sql/entity-per-page.sql
@@ -14,31 +14,20 @@ FROM (
     AVG(executionTime) AS averageExecutionTime
   FROM (
     SELECT
-  domain,
-  canonicalDomain,
-  category,
-  JSON_VALUE(report, '$.requestedUrl') AS pageUrl,
-  NET.HOST(JSON_VALUE(bootupTimeItems, '$.url')) AS observedDomain,
-  JSON_VALUE(bootupTimeItems, '$.url') AS scriptUrl,
-  SAFE_CAST(JSON_VALUE(bootupTimeItems,
-      "$.scripting") AS FLOAT64) AS executionTime
-FROM (
-  SELECT
-    url AS page,
-    report
-  FROM
-    `httparchive.lighthouse.2022_01_01_mobile`
-  ),
-  UNNEST(JSON_QUERY_ARRAY(report,
-      '$.audits.bootup-time.details.items')) AS bootupTimeItems
-  INNER JOIN
-    `lighthouse-infrastructure.third_party_web.2022_01_01`
-  ON NET.HOST(JSON_VALUE(bootupTimeItems, "$.url")) = domain
+      domain,
+      canonicalDomain,
+      JSON_VALUE(report, '$.requestedUrl') AS pageUrl,
+      SAFE_CAST(JSON_VALUE(bootupTimeItems,"$.scripting") AS FLOAT64) AS executionTime
+    FROM (
+      SELECT url AS page, report
+      FROM `httparchive.lighthouse.2022_01_01_mobile`
+    ),
+    UNNEST(JSON_QUERY_ARRAY(report,'$.audits.bootup-time.details.items')) AS bootupTimeItems
+    INNER JOIN `lighthouse-infrastructure.third_party_web.2022_01_01`
+    ON NET.REG_DOMAIN(JSON_VALUE(bootupTimeItems, "$.url")) = domain
   )
   WHERE canonicalDomain IS NOT NULL
   GROUP BY pageUrl, canonicalDomain
 )
-GROUP BY
-  canonicalDomain
-ORDER BY
-  totalOccurrences DESC
+GROUP BY canonicalDomain
+ORDER BY totalOccurrences DESC