This PR is here to expose and suggest a fix to inaccurate computed data when third party runs a script from dynamic subdomain. 

As an example we took `Kameleoon` third party analytic script. We can see that before our change (and on [2024-06 data](https://github.com/patrickhulce/third-party-web/blob/bb8670dd54fcbf34daeace00c452f39031f854df/data/2024-06-01-entity-scripting.json)), it has an `averageExecutionTime` of [`~245ms`](https://github.com/patrickhulce/third-party-web/blob/bb8670dd54fcbf34daeace00c452f39031f854df/data/2024-06-01-entity-scripting.json#L2862) on a total of [`988` websites](https://github.com/patrickhulce/third-party-web/blob/bb8670dd54fcbf34daeace00c452f39031f854df/data/2024-06-01-entity-scripting.json#L2860) which is pretty good but not ranked on README because of its low occurences (below 1000).

What's wrong here is that this third party uses a dynamic subdomain to serve its main script on websites (e.g `9e9soula8o.kameleoon.eu`). Some of these subdomain scripts are kept in final result but analyzing raw `observed-domains` we found a lot that are ignored (~450). 

In this MR we've decided to group and count on registered domain (`NET.REG_DOMAIN`) instead of hostname (`NET.HOST`) to include all subdomains and do not skip "dynamic" subdomains ( as they participate in the average execution time). 

We've also cleaned unused select in query (`COUNT(0) as totalOccurences` nested query).

After updating data with this change, we found that third party using dynamic subdomains are less favored in ranking because of much more accurate `averageExecutionTime` closer to its concurrents. As an example, for Kameleoon, we found that its `totalOccurences` grows-up to `2272`(+1284) meaning it now appears in README ranking (n°63) but with a much more accurate `averageExecutionTime` value of `~714ms` (+469ms). 

As a concurrent comparison we can list:

| Rank | Product | `averageExecutionTime` BEFORE fix | `averageExecutionTime` AFTER fix | Diff |
| ------------- | ------------- | ------------- | ------------- | ------------- |
| 63 | Kameleoon  |  [245ms](https://github.com/patrickhulce/third-party-web/blob/5174773c958268599822f5c5c5248dfbb93e3f5b/data/2024-06-01-entity-scripting.json#L2862) |  714ms | +469ms
| 68 | Optimizely  | [857ms](https://github.com/patrickhulce/third-party-web/blob/5174773c958268599822f5c5c5248dfbb93e3f5b/data/2024-06-01-entity-scripting.json#L1035)  |  861ms | +4ms
| 60 | VWO  | [638ms](https://github.com/patrickhulce/third-party-web/blob/5174773c958268599822f5c5c5248dfbb93e3f5b/data/2024-06-01-entity-scripting.json#L1413)  |  638ms | +0ms
| 55 | ABTasty  | [492ms](https://github.com/patrickhulce/third-party-web/blob/5174773c958268599822f5c5c5248dfbb93e3f5b/data/2024-06-01-entity-scripting.json#L1875)  |  497ms | +5ms


We're pleased to get feedbacks and discuss our demonstration to maybe find a more relevant solution that could prevent this kind of issue without misrepresenting all third-parties.